+++ Dragon-taro/index.html
@@ -1,5 +1,7 @@
 <html>
-    <head></head>
+    <head>
+        <base href="/" />
+    </head>
     <body>
         <div id="root"></div>
         <script src="./static/js/bundle.js"></script>
