+++ Dragon-taro/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "watch": "npx webpack --mode development --watch",
     "build": "npx webpack --mode production",
-    "start": "http-server -o",
+    "start": "webpack-dev-server --hot --inline --progress --colors --history-api-fallback",
     "build-css": "gulp build",
     "watch-css": "gulp watch"
   },
@@ -14,6 +14,7 @@
   "license": "ISC",
   "dependencies": {
     "@babel/polyfill": "^7.0.0",
+    "history": "^4.7.2",
     "oauthio-web": "^0.6.2",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
@@ -33,9 +34,9 @@
     "gulp": "^4.0.0",
     "gulp-postcss": "^8.0.0",
     "gulp-sass": "^4.0.1",
-    "http-server": "^0.11.1",
     "postcss-cssnext": "^3.1.0",
     "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2"
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.9"
   }
 }
