+++ Dragon-taro/src/index.js
@@ -1,13 +1,31 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { HashRouter, Route, Switch } from "react-router-dom";
+import { Router, Route, Switch } from "react-router-dom";
+import createHistory from "history/createBrowserHistory";
+import { createStore, applyMiddleware } from "redux";
+import createSagaMiddleware from "redux-saga";
 import Header from "./containers/HeaderContainer";
 import Home from "./containers/HomeContainer";
 import Gist from "./containers/GistContainer";
 import { Provider } from "react-redux";
-import configureStore from "./store";
 import NewGist from "./containers/NewGistContainer";
 import EditGist from "./containers/EditGistContainer";
+import reducer from "./reducers/reducers";
+import rootSaga from "./sagas/sagas";
+
+const history = createHistory();
+
+function configureStore(initialState) {
+  const sagaMiddleware = createSagaMiddleware();
+  const store = createStore(
+    reducer,
+    initialState,
+    applyMiddleware(sagaMiddleware)
+  );
+
+  sagaMiddleware.run(rootSaga, { history });
+  return store;
+}
 
 class App extends React.Component {
   constructor() {
@@ -16,25 +34,22 @@ class App extends React.Component {
 
   render() {
     return (
-      <HashRouter>
-        <div>
-          <Header />
-          <hr />
-          <Switch>
-            <Route exact path="/" component={Home} />
-            <Route path="/gists/new" component={NewGist} />
-            <Route path="/gists/:id/edit" component={EditGist} />
-            <Route path="/gists/:id" component={Gist} />
-          </Switch>
-        </div>
-      </HashRouter>
+      <Router history={history}>
+        <Provider store={configureStore()}>
+          <div>
+            <Header />
+            <hr />
+            <Switch>
+              <Route exact path="/" component={Home} />
+              <Route path="/gists/new" component={NewGist} />
+              <Route path="/gists/:id/edit" component={EditGist} />
+              <Route path="/gists/:id" component={Gist} />
+            </Switch>
+          </div>
+        </Provider>
+      </Router>
     );
   }
 }
 
-ReactDOM.render(
-  <Provider store={configureStore()}>
-    <App />
-  </Provider>,
-  document.getElementById("root")
-);
+ReactDOM.render(<App />, document.getElementById("root"));
