+++ Dragon-taro/src/sagas/sagas.js
@@ -2,7 +2,7 @@ import { fork } from "redux-saga/effects";
 import oauth from "./oauth";
 import gists from "./gists";
 
-export default function* rootSaga() {
-  yield fork(oauth);
-  yield fork(gists);
+export default function* rootSaga(context) {
+  yield fork(oauth, context);
+  yield fork(gists, context);
 }
