--- Dragon-taro/src/store.js
@@ -1,16 +1,0 @@
-import { createStore, applyMiddleware } from "redux";
-import createSagaMiddleware from "redux-saga";
-import reducer from "./reducers/reducers";
-import rootSaga from "./sagas/sagas";
-
-export default function configureStore(initialState) {
-  const sagaMiddleware = createSagaMiddleware();
-  const store = createStore(
-    reducer,
-    initialState,
-    applyMiddleware(sagaMiddleware)
-  );
-
-  sagaMiddleware.run(rootSaga);
-  return store;
-}
